projects/LitePT/litept/models/modules.py
import sys
from collections import OrderedDict

import spconv.pytorch as spconv
import torch
import torch.nn as nn
import torch_scatter
from addict import Dict

from litept.engines.hooks import HookBase
from litept.models.scatter import argsort, segment_csr, unique
from litept.models.utils.structure import Point


class PointModule(nn.Module):
    r"""PointModule
    placeholder, all module subclass from this will take Point in PointSequential.
    """

    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)


class PointSequential(PointModule):
    r"""A sequential container.
    Modules will be added to it in the order they are passed in the constructor.
    Alternatively, an ordered dict of modules can also be passed in.
    """

    def __init__(self, *args, **kwargs):
        super().__init__()
        if len(args) == 1 and isinstance(args[0], OrderedDict):
            for key, module in args[0].items():
                self.add_module(key, module)
        else:
            for idx, module in enumerate(args):
                self.add_module(str(idx), module)
        for name, module in kwargs.items():
            if sys.version_info < (3, 6):
                raise ValueError("kwargs only supported in py36+")
            if name in self._modules:
                raise ValueError("name exists.")
            self.add_module(name, module)

    def __getitem__(self, idx):
        if not (-len(self) <= idx < len(self)):
            raise IndexError("index {} is out of range".format(idx))
        if idx < 0:
            idx += len(self)
        it = iter(self._modules.values())
        for i in range(idx):
            next(it)
        return next(it)

    def __len__(self):
        return len(self._modules)

    def add(self, module, name=None):
        if name is None:
            name = str(len(self._modules))
            if name in self._modules:
                raise KeyError("name exists")
        self.add_module(name, module)

    def forward(self, input):
        for k, module in self._modules.items():
            # Point module
            if isinstance(module, PointModule):
                input = module(input)
            # Spconv module
            elif spconv.modules.is_spconv_module(module):
                if isinstance(input, Point):
                    input.sparse_conv_feat = module(input.sparse_conv_feat)
                    input.feat = input.sparse_conv_feat.features
                else:
                    input = module(input)
            # PyTorch module
            else:
                if isinstance(input, Point):
                    input.feat = module(input.feat)
                    if "sparse_conv_feat" in input.keys():
                        input.sparse_conv_feat = input.sparse_conv_feat.replace_feature(input.feat)
                elif isinstance(input, spconv.SparseConvTensor):
                    if input.indices.shape[0] != 0:
                        input = input.replace_feature(module(input.features))
                else:
                    input = module(input)
        return input


class PointModel(PointModule, HookBase):
    r"""PointModel
    placeholder, PointModel can be customized as a Pointcept hook.
    """

    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)


class GridPooling(PointModule):
    def __init__(
        self,
        in_channels,
        out_channels,
        stride=2,
        norm_layer=None,
        act_layer=None,
        reduce="max",
        shuffle_orders=True,
        traceable=True,  # record parent and cluster
        re_serialization=False,
        serialization_order="z",
    ):
        super().__init__()
        self.in_channels = in_channels
        self.out_channels = out_channels

        self.stride = stride
        assert reduce in ["sum", "mean", "min", "max"]
        self.reduce = reduce
        self.shuffle_orders = shuffle_orders
        self.traceable = traceable

        self.proj = nn.Linear(in_channels, out_channels)
        if norm_layer is not None:
            self.norm = PointSequential(norm_layer(out_channels))
        if act_layer is not None:
            self.act = PointSequential(act_layer())

        self.re_serialization = re_serialization
        self.serialization_order = serialization_order

    def forward(self, point: Point):
        if "grid_coord" in point.keys():
            grid_coord = point.grid_coord
        elif {"coord", "grid_size"}.issubset(point.keys()):
            grid_coord = torch.div(
                point.coord - point.coord.min(0)[0],
                point.grid_size,
                rounding_mode="trunc",
            ).int()
        else:
            raise AssertionError(
                "[gird_coord] or [coord, grid_size] should be include in the Point"
            )
        grid_coord = torch.div(grid_coord, self.stride, rounding_mode="trunc")

        # Pack batch id and coordinates into a single 1D key for clustering.
        # NOTE(original): grid_coord = torch.bitwise_or(grid_coord, point.batch.view(-1, 1) << 48)
        grid_coord_i = grid_coord.to(torch.int64)
        gx = grid_coord_i[:, 0]
        gy = grid_coord_i[:, 1]
        gz = grid_coord_i[:, 2]
        b = point.batch.to(torch.int64)
        packed = gx + gy * (1 << 16) + gz * (1 << 32) + b * (1 << 48)

        if torch.onnx.is_in_onnx_export():
            unique_keys, cluster, counts, _ = unique(packed)
            # Sort points by cluster id (inverse_indices) to make them contiguous per cluster.
            indices = argsort(cluster)
        else:
            unique_keys, cluster, counts = torch.unique(
                packed,
                sorted=True,
                return_inverse=True,
                return_counts=True,
            )
            # indices of point sorted by cluster, for torch_scatter.segment_csr
            _, indices = torch.sort(cluster)

        # Unpack to (M, 3) grid coords (drop batch component)
        # NOTE(original): grid_coord = torch.bitwise_and(grid_coord, ((1 << 48) - 1))
        key_wo_batch = torch.remainder(unique_keys, (1 << 48))
        gx = torch.remainder(key_wo_batch, (1 << 16))
        gy = torch.remainder(torch.div(key_wo_batch, (1 << 16), rounding_mode="trunc"), (1 << 16))
        gz = torch.remainder(torch.div(key_wo_batch, (1 << 32), rounding_mode="trunc"), (1 << 16))
        grid_coord = torch.stack([gx, gy, gz], dim=1).to(grid_coord_i.dtype)

        # index pointer for sorted point, for torch_scatter.segment_csr
        idx_ptr = torch.cat([counts.new_zeros(1), torch.cumsum(counts, dim=0)])
        # head_indices of each cluster, for reduce attr e.g. code, batch
        head_indices = indices[idx_ptr[:-1]]

        if not torch.onnx.is_in_onnx_export():
            scatter_feat = torch_scatter.segment_csr(
                self.proj(point.feat)[indices], idx_ptr, reduce=self.reduce
            )
            scatter_coord = torch_scatter.segment_csr(point.coord[indices], idx_ptr, reduce="mean")
        else:
            scatter_feat = segment_csr(self.proj(point.feat)[indices], idx_ptr, self.reduce)
            scatter_coord = segment_csr(point.coord[indices], idx_ptr, "mean")

        point_dict = Dict(
            feat=scatter_feat,
            coord=scatter_coord,
            grid_coord=grid_coord,
            batch=point.batch[head_indices],
        )

        if "origin_coord" in point.keys():
            if not torch.onnx.is_in_onnx_export():
                point_dict["origin_coord"] = torch_scatter.segment_csr(
                    point.origin_coord[indices], idx_ptr, reduce="mean"
                )
            else:
                point_dict["origin_coord"] = segment_csr(
                    point.origin_coord[indices], idx_ptr, "mean"
                )
        if "condition" in point.keys():
            point_dict["condition"] = point.condition
        if "context" in point.keys():
            point_dict["context"] = point.context
        if "name" in point.keys():
            point_dict["name"] = point.name
        if "split" in point.keys():
            point_dict["split"] = point.split
        if "color" in point.keys():
            if not torch.onnx.is_in_onnx_export():
                point_dict["color"] = torch_scatter.segment_csr(
                    point.color[indices], idx_ptr, reduce="mean"
                )
            else:
                point_dict["color"] = segment_csr(point.color[indices], idx_ptr, "mean")
        if "grid_size" in point.keys():
            point_dict["grid_size"] = point.grid_size * self.stride
        if "mask" in point.keys():
            if not torch.onnx.is_in_onnx_export():
                point_dict["mask"] = (
                    torch_scatter.segment_csr(point.mask[indices].float(), idx_ptr, reduce="mean")
                    > 0.5
                )
            else:
                point_dict["mask"] = segment_csr(point.mask[indices].float(), idx_ptr, "mean") > 0.5

        if self.traceable:
            point_dict["pooling_inverse"] = cluster
            point_dict["pooling_parent"] = point
        point = Point(point_dict)
        if self.norm is not None:
            point = self.norm(point)
        if self.act is not None:
            point = self.act(point)

        if self.re_serialization:
            point.serialization(order=self.serialization_order, shuffle_orders=self.shuffle_orders)
        point.sparsify()
        return point


class GridUnpooling(PointModule):
    def __init__(
        self,
        in_channels,
        skip_channels,
        out_channels,
        norm_layer=None,
        act_layer=None,
        traceable=False,  # record parent and cluster
    ):
        super().__init__()
        self.proj = PointSequential(nn.Linear(in_channels, out_channels))
        self.proj_skip = PointSequential(nn.Linear(skip_channels, out_channels))

        if norm_layer is not None:
            self.proj.add(norm_layer(out_channels))
            self.proj_skip.add(norm_layer(out_channels))

        if act_layer is not None:
            self.proj.add(act_layer())
            self.proj_skip.add(act_layer())

        self.traceable = traceable

    def forward(self, point):
        assert "pooling_parent" in point.keys()
        assert "pooling_inverse" in point.keys()
        parent = point.pop("pooling_parent")
        inverse = point.pooling_inverse
        feat = point.feat

        parent = self.proj_skip(parent)
        parent.feat = parent.feat + self.proj(point).feat[inverse]
        parent.sparse_conv_feat = parent.sparse_conv_feat.replace_feature(parent.feat)

        if self.traceable:
            point.feat = feat
            parent["unpooling_parent"] = point
            parent["unpooling_inverse"] = inverse
        return parent


class Embedding(PointModule):
    def __init__(
        self,
        in_channels,
        embed_channels,
        norm_layer=None,
        act_layer=None,
        export_mode=False,
    ):
        super().__init__()
        self.in_channels = in_channels
        self.embed_channels = embed_channels

        if export_mode:
            from SparseConvolution.sparse_conv import SubMConv3d
        else:
            from spconv.pytorch import SubMConv3d

        self.stem = PointSequential(
            conv=SubMConv3d(
                in_channels,
                embed_channels,
                kernel_size=5,
                padding=1,
                bias=False,
                indice_key="stem",
            )
        )
        if norm_layer is not None:
            self.stem.add(norm_layer(embed_channels), name="norm")
        if act_layer is not None:
            self.stem.add(act_layer(), name="act")

    def forward(self, point: Point):
        point = self.stem(point)
        return point
