projects/LitePT/pyproject.toml
[project]
name = "litept"
version = "0.1.0"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
  "addict>=2.4.0",
  "colorhash>=2.1.0",
  "einops>=0.8.1",
  "flash-attn",
  "h5py>=3.15.1",
  "plyfile>=1.1.3",
  "pointrope",
  "pyyaml>=6.0.3",
  "scipy>=1.15.3",
  "sharedarray>=3.2.4",
  "spconv-cu126>=2.3.8",
  "tensorboard>=2.20.0",
  "tensorboardx>=2.6.4",
  "termcolor>=3.3.0",
  "timm>=1.0.22",
  "torch==2.8.0",
  "torchvision==0.23.0",
  "torchmetrics>=1.8.2",
  "torch-cluster>=1.6.3",
  "torch-geometric>=2.7.0",
  "torch-scatter>=2.1.2",
  "torch-sparse>=0.6.18",
  "wandb>=0.23.1",
  "yapf>=0.43.0",
  "nuscenes-devkit>=1.2.0",
  "open3d>=0.19.0",
  "onnx>=1.20.0",
]

[project.optional-dependencies]
insseg = ["pointgroup-ops", "pointops"] # for instance segmentation

[tool.uv]
find-links = [
  "https://data.pyg.org/whl/torch-2.8.0+cu128.html", # for torch-* libraries
]
no-build-isolation-package = [
  "flash-attn",
  "torch-cluster",
  "torch-geometric",
  "torch-scatter",
  "torch-sparse",
  "pointops",
  "pointrope",
  "pointgroup-ops",
]

[tool.uv.extra-build-dependencies]
pointrope = [{ requirement = "torch", match-runtime = true }]

[tool.uv.sources]
torch = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux'" }]
torchvision = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux'" }]
flash-attn = { url = "https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.3/flash_attn-2.7.3+cu12torch2.8cxx11abiFALSE-cp310-cp310-linux_x86_64.whl" }
pointrope = { workspace = true }
pointops = { workspace = true }
pointgroup-ops = { workspace = true }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.workspace]
members = ["libs/pointgroup_ops", "libs/pointops", "libs/pointrope"]

[tool.ruff]
line-length = 100
